refactor chat page into components
